index planning: implement cost model predicate (#12197)
#### What this PR does

 is merged and vendored


This PR introduces the plan predicate. The predicate is a label matcher.
The predicate has a selectivity, cost for applying it to a single value,
and cardinality. All of these are estimations. The estimations use the
changes from https://github.com/grafana/mimir-prometheus/pull/940

The code right now is not used, but will be when
https://github.com/grafana/mimir/issues/11919 is finished


### Testing
This PR also introduces testing for the 3 properties of the predicate. 

The way I decided to approach testing was inspired by how CockroachDB
approach testing their query cost model. They have query plans'
estimates costs hard-coded (for example [unit tests for `SELECT`
queries](https://github.com/cockroachdb/cockroach/blob/b8c2405930718b735bf009c802798b4918b66631/pkg/sql/opt/xform/testdata/coster/select);
benchmarking for TPCH merge-join
[plan](https://github.com/cockroachdb/cockroach/blob/b8c2405930718b735bf009c802798b4918b66631/pkg/sql/opt/opbench/opbench_test.go#L359-L407)
and its [cost in
CSV](https://github.com/cockroachdb/cockroach/blob/b8c2405930718b735bf009c802798b4918b66631/pkg/sql/opt/opbench/testdata/tpch-merge-join.csv)).

1. Hard-coded mock statistics

This was easy to generate, it's easy to reason about and to write
assertions against. An alternative is to have a generated statistics
based on some cardinality code. Another alternative is to dump the
cardinality from a real TSDB and use that.
   
The latter would be most realistic, but I'm unsure about confidentiality
and about navigating this cardinality dump (i.e. writing new test cases
against it).
3. CSV file instead of typical golang table tests.

The reason that they're CSV files is so that when we improve the model,
we can toggle a boolean in the test and then rewrite the files with the
new estimates (inspired by cockroach's benchmarks). For this, I've also
added a CSV mapper to help with mapping to and from CSV files.
   
An alternative I tried is to only assert on the relative
costs/cardinalities of different matchers. This was hard to reason about
when there are more than a few matchers together in the same test case.
If the test cases were few, then there were just a lot of test cases.
   
   

#### Which issue(s) this PR fixes or relates to

related to https://github.com/grafana/mimir/issues/11919

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>