ingest-storage: Split very large RW2.0 requests into multiple Kafka records (#12166)
#### What this PR does

This is the record format V2 equivalent of
https://github.com/grafana/mimir/pull/8077. Kafka has a limit in payload
size, here we break down extremely large requests into multiple records
to work around it.

RW2 carries some complexity due to the symbols table. When splitting a
request, we have to split the symbols as well. It's not reasonable to
use all the symbols on each request, because symbols still dominate the
overall request size.

This PR uses a similar bin-packing algorithm to RW1, but it also
re-symbolizes each Timeseries as we move along, using a separate symbols
table per sub-request. That means each Timeseries can re-use symbols
from other timeseries in the same request as they're added, up until the
point where the request is close to the max.

To determine when to split a request, we calculate an upper bound of the
growth that the next timeseries will cause without actually processing
any symbols yet. If it's possible that this timeseries can push things
over the limit, we split it (preferring splitting early vs breaching the
limit).

Addresses this comment:
https://github.com/grafana/mimir/pull/12060#discussion_r2221466520

#### Which issue(s) this PR fixes or relates to

contrib https://github.com/grafana/mimir-squad/issues/2253

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.