MQE: change log message to disambiguate from query-frontend query stats log messages (#12251)
#### What this PR does

The query-frontend emits a `query stats` log message after each query
request is completed.

MQE also emits a `query stats` log message after evaluation is
completed.

With MQE now running in query-frontends, this is potentially confusing,
so I've changed the log message from MQE to `evaluation stats` to make
it unambiguous.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/pull/11417

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.