Refactor query response Protobuf definition (#12252)
#### What this PR does

This PR refactors the definition of `QueryResponse` to allow use of the
error type enum in other messages. (This other use is something I'll
introduce in a follow-up PR.)

This required prefixing the names of the `ErrorCause` values, as both
`QueryResponse` and `ErrorCause` have a "bad data" value, and enum
values are not scoped by their enum name (see
https://protobuf.dev/programming-guides/style/#enum-value-prefixing).

For consistency, I've also promoted the `QueryStatus` enum to top level
as well.

While these changes affect the Go code that consume these messages, it
does not change the wire format in any way, so there are no backwards
compatibility issues due to these changes.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.