Remove instant query splitting feature (#12267)
#### What this PR does

This PR removes the experimental instant query splitting feature. As
discussed in #12253, there are a couple of correctness issues with the
existing implementation and we're working on a new implementation in
MQE.

#### Which issue(s) this PR fixes or relates to

Resolves https://github.com/grafana/mimir/issues/12253

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [x]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.