Pass ctx to Map and MapExpr for ASTMapper (#12238)
#### What this PR does

Pass ctx to `Map` and `MapExpr` for `ASTMapper`, instead of storing it
in the structs, so that we don't have to recreate the mapper when we
only want to change ctx, and handle context cancellation in the
`ASTMapper` so implementations don't have to. Suggested by @charleskorn
in
[here](https://github.com/grafana/mimir/pull/11971#discussion_r2196320853).

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.