[HELM] Add labels configuration for gateway ingress (#11964)
This update introduces a new `labels` field under the `gateway.ingress`
section in the `values.yaml` file, allowing users to specify custom
labels for the Ingress resource. The corresponding template in
`gateway-ingress.yaml` has been updated to include these labels in the
metadata.

- Added `labels: {}` to `values.yaml`
- Updated `gateway-ingress.yaml` to render labels if provided

This enhancement improves the flexibility of the Ingress configuration
for users deploying the Mimir distributed system.

Checklist

- [X]  Tests updated
- [X]  Documentation added
- [X] CHANGELOG.md updated - the order of entries should be [CHANGE],
[FEATURE], [ENHANCEMENT], [BUGFIX]

---------

Signed-off-by: unkls <benjamin.leveque97@gmail.com>
Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>
Co-authored-by: Vladimir Varankin <vladimir.varankin@grafana.com>