Remove unsafe from mimirpb.CopyLabels (#12292)
#### What this PR does

Part of a bigger effort to contain the unsafety within mimirpb.

`CopyLabels` can be implemented in a safe manner with the same memory
optimization by using a StringBuilder rather than a `[]byte` and
`yoloString`.

Note that this code only runs when `stringlabels` build tag isn't set.
Official Mimir builds do set `stringlabels` so this code won't actually
run.

#### Which issue(s) this PR fixes or relates to


https://grafana.com/blog/2025/08/03/grafana-cloud-metrics-memory-corruption-issue-resolved/

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.