Alertmanager: Refactor createUsableGrafanaConfig (#12289)
This PR refactors the `createUsableGrafanaConfig` function.

Three changes:
- Rename to `amConfigFromGrafanaConfig`: It's not clear what "usable"
means. The name `amConfigFromGrafanaConfig` implies we're creating an
`amConfig` from a Grafana configuration.
- Make it a method in the `MultitenantAlertmanager`: we can use the
logger and fallback configs in the struct instead of always passing them
as arguments
- Remove logger and config arguments: we're always using the same config
and loggers, it's unnecessary for the caller to always pass these

Part of https://github.com/grafana/mimir/pull/12272