mimir-mixin: relax the guard in IngesterHasUnshippedBlocks when block-builder is running (#12319)
#### What this PR does

This is an alternative fix for the issue,
https://github.com/grafana/mimir/pull/12314 wanted to address
(_providing I understood the motivation for later_).

We observed a couple of cases, where the block-builder deployment was
restarted right around the end of its hourly cycle. This reset its
`successful_compact_and_upload_timestamp_seconds`, and delays the next
update of the metric, until the updated block-builders catch up. In
turn, this caused the `MimirIngesterHasUnshippedBlocks` to fire, because
the mentioned metric wasn't set.

This PR updates the `MimirIngesterHasUnshippedBlocks` (and related)
alerts, relaxing the `unless ...` hand, so it looked it for the past N
minutes, when it checked for block-builder's
`successful_compact_and_upload_timestamp_seconds`.

---------

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>