Warn about unsafe mutable labels in type names and doc strings (#12271)
#### What this PR does

* Creates aliases for `string` (`unsafeMutableString`) and
`mimirpb.LabelAdapter` (`UnsafeMutableLabel`) to try and make it more
obvious when it's not safe to shallow-copy then, as it usually is with
strings/structs of strings.
* Add notices in the doc comments of `PreallocTimeseries` and
descendents about avoiding shallow copies.

#### Checklist

- [x] Tests updated.
- [x] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.