TestValidateLabels: run finalChecks from the cleanup (#12354)
Instead of doing the unnecessary atomic logic, we can just register a
cleanup function: that runs after all subtests are finished.

But the main reason why I'm changing this is because the test that
should fail is the outer one, but instead with the atomic approach a
random subtest fails every time, depending on their order of the
execution.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>