mimirpb: Associate metadata to series when possible during RW2 conversion (#12342)
#### What this PR does

RW2 attaches metric metadata to series, where RW1 carries metadata
separately.
Previously, we took a fast path, where we always generate new series
just to carry each metadata.
This PR changes the conversion to metadata into the corresponding
series, if both are present in the same request.

The main benefit is on requests that are already RW2 (and tend to carry
a mixture of series and metadata). We squeeze some additional bytes out
of the new format by not repeating the label name references two times
per metric family, plus dropping extra `repeated TimeSeriesRW2`
separators in protobuf. It does not change data-only or metadata-only
pushes.

We now do extra work in the distributor to match these elements up.
Previously we theorized this wasn't worth it, but now after analyzing
typical distributor profile data, I expect the extra work here to not be
as impactful on perf or TCO as once thought. The upside is we keep the
record size absolutely minimized, as clients shift over to send RW2 push
requests.

#### Which issue(s) this PR fixes or relates to

contrib: https://github.com/grafana/mimir-squad/issues/2253

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.