Make MQE the default query engine in query-frontends (#12361)
#### What this PR does

We've been running query-frontends with MQE enabled in production at
Grafana Labs for quite a while now, and have had no issues.

So in this PR I'm making MQE the default. It's still possible to switch
back to Prometheus' engine by setting
`-query-frontend.query-engine=prometheus`.

#### Which issue(s) this PR fixes or relates to

Resolves https://github.com/grafana/mimir/issues/11882

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.