Distributor: Auto-deduce OTel translation strategy in OTLP endpoint if unset (#12369)
#### What this PR does

Modify validation code to not configure default OTel metric and label
name translation strategy, but instead leave that to the distributor's
OTLP endpoint logic. This is because we found a bug connected to
automatically deducing the translation strategy within
`Limits.Validate`, and it might as well be done at the call site (in the
OTLP endpoint).

#### Which issue(s) this PR fixes or relates to

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Arve Knudsen <arve.knudsen@gmail.com>