refactor: make query-scheduler a required component (#12200)
This change removes the ability to use the v1 query-frontend which used
an embedded version of the query-scheduler. Instead, v2 of the
query-frontend is now required which requires a stand-alone deployment
of the query-scheduler. Note that this change does _not_ do any clean up
of the query-frontend, it only removes v1.

Part of https://github.com/grafana/mimir/issues/11884