MQE: shift responsibilities of `Materializer` (#12377)
#### What this PR does

This PR changes the responsibilities of MQE's `Materializer` type.

Previously, `Materializer.ConvertNodeToOperator` would materialize all
of the given node's children and pass it to `Node.OperatorFactory`, and
`OperatorFactory` would have no access to the `Materializer`. With the
existing set of query plan nodes, this was not an issue, but it is
problematic for some scenarios, for example:

* for remote execution, there's no point materializing operators for the
parts of the query plan that will be executed remotely
* for caching and splitting, there's also no point materializing
operators for cached results, but we'll then need to materialize the
child operator if there is a cache miss, or only materialize the child
operator for a subset of the time range

This PR also moves `Materializer` to the `planning` package to prevent
an import cycle (and seems like a more natural fit for there anyway).

Depending on which PR merges first, either this PR or
https://github.com/grafana/mimir/pull/12302 will need changes to account
for the changes in the other.

This PR has no user-visible impact and so I've chosen not to add a
changelog entry.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.