Remote execution of query plans in queriers (#12302)
#### What this PR does

This PR adds support for remote execution of query plans in queriers.

Query-frontends can now send a Protobuf message (instead of a smuggled
HTTP request) to queriers via query-schedulers, and queriers will
respond with a stream of results. This has been implemented in a way
where we can add other kinds of query requests and responses in the
future, with the goal of eventually moving away from smuggling HTTP
requests in Protobuf messages.

In the case of queries, query-frontends can now send a query plan and
request evaluation of a particular node in the plan, and queriers will
respond with results for that node.

I've also added a HTTP endpoint to queriers for debugging purposes
(`/api/v1/evaluate`), and a CLI tool to exercise this
(`tools/evaluate-query-plan`). I plan to remove this in a subsequent PR,
once I've added the functionality to query-frontends to send these
requests to queriers.

Not included in this PR:
* batching of responses to query-frontends (each series' data is sent as
its own Protobuf message for now)
* actually sending requests from query-frontends
* metrics and tracing for requests to queriers

There are a number of other items I'll handle in follow up PRs (eg.
propagating strong consistency information and dealing with query stats)
- my goal with this PR is to get the general structure in place and
start getting feedback on the approach before I invest too much more
time in this.

I've chosen not to add a changelog entry given this code is not actively
used at present, but open to other opinions on this.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.