MQE: separate responsibilities of `Node` (#12390)
#### What this PR does

This PR refactors the responsibilities of the `planning.Node` type to
remove responsibility for creating `OperatorFactory` instances.

Instead, this responsibility has been moved to the newly-introduced
`NodeMaterializer` interface.

This will allow us to implement new kinds of nodes that can't be
materialized with only values available from `OperatorParameters`, and
where it doesn't make sense to extend `OperatorParameters` to include
these values. For example, a caching operator might need cache
connection information, and a remote execution node might need access to
a client for remote executors, but these things are specific to only
these kinds of nodes, so it doesn't make sense to have the engine itself
depend on these things.

From a code organisation perspective, it also means the
`streamingpromql` package won't need to depend on a bunch of other
packages that may introduce circular dependencies (eg. for remote
execution, we would likely need to reference query-frontend packages,
but the query-frontend also needs to reference the `streamingpromql`
package).

I've not added a changelog entry given this should have no user-visible
impact.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.