MQE: query-frontend refactoring to support remote execution of query plans in queriers (#12394)
#### What this PR does

This PR refactors parts of the query-frontend to make the introduction
of remote execution of query plans in queriers easier.

I've tried to make each commit self-contained, and would recommend
reviewing each separately.

This PR does not need a changelog entry as it makes no user-visible
changes.

#### Which issue(s) this PR fixes or relates to

Related to #12302

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.