MQE: add the ability to compute the queried time range for a query plan node (#12396)
#### What this PR does

This PR extends the `planning.Node` interface to add a
`QueriedTimeRange` method.

This is needed to allow computing the queried components (ingesters or
store-gateways or both) when remotely evaluating a query plan, so that
query-schedulers can prioritise requests appropriately when one is under
load. See https://github.com/grafana/mimir/pull/6772 for more details of
this.

This has no user-visible impact, so I have not added a changelog entry.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/pull/6772

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.