MQE: run fewer variations in `TestCompareVariousMixedMetricsAggregations`, to dramatically reduce test run times (#12409)
#### What this PR does

This halves the time taken to run all tests in `pkg/streamingpromql` on
my machine, from ~60s to ~30s. On CI, the time drops from ~14m to ~7m.

Given `pkg/streamingpromql` is one of the slowest test suites to run,
this should have a noticeable impact on CI run times after #12392 is
merged, where unit tests become the critical path for CI run times.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.