Multiple Components: Include vector(0) in Memory HPA Metric (#12415)
#### What this PR does

This PR follows https://github.com/grafana/mimir/pull/12386. The change
to the CPU and memory HPA metrics must be re-worked to avoid breaking
scaling when there are no OOMKilled pods. We previously
[reverted](https://github.com/grafana/mimir/pull/12400) the CPU HPA and
partially reverted the memory change.This PR fully restores the memory
scaling metric (see comment:
[here](https://github.com/grafana/mimir/pull/12386/files#r2276681205)).

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.