MQE: introduce `Node.ExpressionPosition()` method (#12410)
#### What this PR does

This PR introduces an `ExpressionPosition` method to `planning.Node`.

This is necessary to support remote execution, because the remote
execution operator needs to be able to provide an expression position
based on the node it will evaluate.

I have not added a changelog entry as this should have no user-visible
impact.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.