Expose rollout_operator_enabled into $._config (#12419)
#### What this PR does

Expose the rollout_operator_enabled variable into $._config so it can be
referenced externally.

#### Which issue(s) this PR fixes or relates to

Exposing this variable in the config will allow for better conditions in
deployment-tools libsonnet overrides. eg
https://github.com/grafana/deployment_tools/pull/324750.

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.