Rollout dashboard: make panels higher (#12429)
#### What this PR does

The number of components in a Mimir deployment grows over time, and it's
becoming challenging to visualize them in the Rollout progress panel, so
I made it higher, as well ass the "Pods count per version" one, moving
the "Unhealthy pods" one to the corner.

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>