chore: Update default dynamic replication multiple (#12433)
Update the default value of the dynamic replication multiple in
store-gateways to `5` to match what we've found to be a good balance in
Grafana Cloud. Using a replication factor of 15 (3*5) is a good balance
between keeping additional memory and disk usage minimal but still
making a noticeable difference in the spread of CPU usage between
replicas.