MQE: don't panic if remote execution attempted on querier with MQE disabled (#12438)
#### What this PR does

This PR fixes a potential issue with
https://github.com/grafana/mimir/pull/12302.

If a query-frontend requests remote evaluation of a query plan on a
querier where MQE is not enabled, then the querier could panic.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/pull/12302

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.