10638 - Fuzz tests for MQE (#12121)
#### What this PR does

This PR contains additional unit tests and Go Fuzz tests which attempt
to validate the Mimir Query Engine.

The tests seek to use a range of query strings to confirm that the Mimir
Query Engine responds in the same manner as the Prometheus Query Engine.
This includes validating both a result, result annotation and/or errors.

There are some expected / accepted scenarios where an exact match does
not occur. These are detailed and accommodated in the test code.

The Fuzz tests include a range of seed / corpus entries which enables
the test suite to be run as a normal unit test. Alternatively, the
specific Fuzz testing function can be invoked using the go test
-fuzz=<func>.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/issues/10638

Fixes #10638

#### Checklist

- [x] Tests updated.
- [x] Documentation added.
- [ ] `CHANGELOG.md` updated
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Andrew Hall <andrew.hall@grafana.com>